refactor: 절대 경로 통일
