--- src/components/common/index.ts
@@ -1,5 +1,0 @@
-export { default as Title } from './Title/Title';
-export { default as Label } from './Label/Label';
-export { default as Input } from './Input/Input';
-export { default as Spacing } from './Spacing/Spacing';
-export { default as ErrorMessage } from './ErrorMessage/ErrorMessage';
