+++ src/stories/components/common/Label.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Label from '../../../components/common/Label/Label';
+import { Label } from '@/components';
 
 const meta = {
   title: 'Label',
